+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 ### Changed
 - (maint) Update vmpooler URL
 - (VANAGON-167) Update tests to remove redhat 5
+- (maint) When using `vanagon list` with the --platforms or --projects flags, no longer print the "-Platforms" or "-Projects" header. 
 
 ### Removed
 - (VANAGON-166) Do not undefine `__debug_package` on EL 8 FIPS